requirements.txt
# local package
#-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
numpy
torch
torchvision
torchaudio
pillow
matplotlib
seaborn
scikit-learn
albumentations
opencv-python
hydra-core
pytest
black
isort
# only for notebooks
ipykernel
jupyterlab
notebook
tqdm